Moving all reusable components tests up a folder for prod build.
