+++ src/components/__tests__/Fieldset-test.tsx
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import * as Enzyme from "enzyme";
 import * as React from "react";
-import Fieldset from "../Fieldset";
+import Fieldset from "../reusables/Fieldset";
 
 describe("Fieldset", () => {
   let wrapper: Enzyme.ShallowWrapper<any, {}>;
