+++ src/components/__tests__/Tabs-test.tsx
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import * as Enzyme from "enzyme";
 import * as React from "react";
-import Tabs from "../Tabs";
+import Tabs from "../reusables/Tabs";
 
 describe("Tabs", () => {
   let wrapper;
